:truck: Rename import for plugin type
