+++ back-office/src/plugins/index.js
@@ -1,7 +1,7 @@
 // @flow
 export * from './props'
 export { default as PluginList } from './PluginList'
-export * from './plugins.type'
+export * from './type'
 export { default as reducer } from './reducer'
 export * from './actions'
 export * from './selectors'
