+++ back-office/src/plugins/reducer.js
@@ -1,7 +1,7 @@
 // @flow
 import { range } from 'lodash'
 import type { Reducer } from 'redux'
-import type { State } from './plugins.type'
+import type { State } from './type'
 import type { Action } from '../store/types'
 
 const intialState = range(20).map(i => ({
